+++ src/controller.h
@@ -14,7 +14,7 @@ typedef struct jackpifm_controller_t jackpifm_controller_t;
 /* jackpifm_controller_new: create new sample rate controller */
 jackpifm_controller_t *jackpifm_controller_new(
   double static_resample_factor,
-  size_t target_delay,
+  double target_delay,
   size_t smooth_size,
   int catch_factor,
   int catch_factor2,
